Preliminary GUI as Discussed this wkend, 1/14
I created this GUI file to match the visual of what we discussed at our meeting this weekend. Since I am unable to upload the file, I have instead created this file source code to let everyone see what it is/how it's made, and hopefully we can pull/branch off of this to add modifications over time as needed. --Beth