Added new features
Rearranged the GUI layout - it's not perfect, but it's a lot closer to what we were talking about. Added time stamp function which is called whenever either the add, edit, or remove buttons are pressed. Currently the function only returns the current time as a time stamp, but I am unsure of what to actually DO with it since that is something that the database wants. There is a bit of dead space in the middle between the combobox and the buttons that I would like to get rid of, I am researching on how. I think it has to do with my layouts, but we will see. I also plan to pad with a little space between the text areas, but for now I just wanted to throw this up and see what you guys thought.